# prevent the utf8 string from being collected by the garbage collector in forder()