--- a/inst/tests/tests.Rraw
+++ b/inst/tests/tests.Rraw
@@ -11519,9 +11519,9 @@ if (identical(enc2utf8(enc2native(utf8_strings)), utf8_strings)) {
   test(1891.3, DT[, uniqueN(x)], 5L)
 
   DT <- data.table(x = mixed_strings, y = c(native_strings, utf8_strings), z = 1)
-  test(1881.4, nrow(DT[, .N, by = .(z, x, y)]), 5L)
-  test(1881.5, nrow(DT[, .N, by = .(y, x, z)]), 5L)
-  test(1881.6, nrow(DT[, .N, by = .(y, z, x)]), 5L)
+  test(1891.4, nrow(DT[, .N, by = .(z, x, y)]), 5L)
+  test(1891.5, nrow(DT[, .N, by = .(y, x, z)]), 5L)
+  test(1891.6, nrow(DT[, .N, by = .(y, z, x)]), 5L)
 }