Closes #2674
This PR replaced PR #2675 (see comments there). 

If somehow the garbage collector was triggered during sorting (like there're millions of non-ASCII characters), it leads to the collapse of `data.table` (see #2674 for details) **because it assumes there're converted UTF-8 chars in the global string pool.** This PR tries to fix this issue. 

In addition, it brings performance enhancement in the case of millions non-ASCII chars, because now it only needs to be converted to UTF-8 once. **Before this PR, the strings need to be converted twice in ` csort_pre()` and `csort()` respectively**, which may be a big issue for a large character vector (for example, on my computer, `enc2utf8()` takes about 20s for a 1e7 length Chinese character).

# TODO
- [x] a new test
- [x] entries in NEWS.md
- [x] `by = .(x, y)` should return the same row as `by = .(y, x)` (see comment below)
- [x] should not affect the performance if all the strings are ASCII or encoded in UTF-8.