@shrektan 2018-03-16
@mattdowle It's the first time I realize that the `radixsort` in R is the same as in `data.table`. Thanks for the great work! I don't think I've encountered any issue before so it should be a problem of data.table-1.10.5 only. 

However, regarding to https://github.com/Rdatatable/data.table/blob/master/src/forder.c#L1227 , are you sure that it's ok because the `xd` here will not be converted to `UTF-8`?

About the performance, it gets significantly improved when there're lots on non-ASCII chars but yes, it becomes a little slower if all the strings are ASCII. See the tests on my computers:

```r
library(data.table)
nonascii_string <- function(n, utf8 = TRUE) {
  x <- c("公允价值变动损益", "红利收入", "价差收入", "其他业务支出", "资产减值损失")
  if (isTRUE(utf8)) x <- enc2utf8(x)
  sample(x, n, TRUE)
}

# test1
tmp <- data.table(x = sample(letters, 1e8, TRUE))
system.time(setkey(tmp, x)) 
# test2
tmp <- data.table(x = sample(letters, 1e8, TRUE), y = sample(letters, 1e8, TRUE))
system.time(setkey(tmp, y, x)) 
# test3
tmp <- data.table(x = nonascii_string(1e7))
system.time(setkey(tmp, x)) 
# test4
tmp <- data.table(x = nonascii_string(1e7), y = nonascii_string(1e7))
system.time(setkey(tmp, y, x)) 
# test5
tmp <- data.table(x = nonascii_string(1e5, FALSE))
system.time(setkey(tmp, x)) 
```
version | test1 | test2 | test3 | test4 | test5
---------|------|-------|-------|-------| -----
v1.9.6  | 2.0s | 6.13s | 0.14s | 0.41s | 0.00s
master  | 2.0s | 6.16s | 0.19s | 0.39s | 0.47s (if not fail)
this version  | 2.50s | 6.94s | 0.18s | 0.39s | 0.14s