@shrektan 2018-03-17
@mattdowle I've pushed more commits to fix the cases on grouping (a.k.a, `DT[, j, by = .(x, y)]`) and the overheads for all ASCII/UTF8 columns.