@mattdowle 2018-03-29
Thanks for all this!  It's looking great to me.  According to codecov, `need2utf8` always returns false. Is there a way to add a test to cover those 4 lines?