@shrektan 2018-03-30
I guess adding an example uses the `Latin-1` encoding may work. I'll submit one later.